fix: Zoom reference window.
